requirements.txt
# Core ML
ultralytics>=8.4.0
torch>=2.1.0
torchvision>=0.16.0

# Custom augmentation
albumentations>=1.4.0
opencv-python-headless>=4.8.0

# Reporting
matplotlib>=3.8.0
pandas>=2.1.0
seaborn>=0.13.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
